pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["tmvfb <darestade@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^2.2.3"
python-dotenv = "^0.20.0"
gunicorn = "^20.1.0"
flake8 = "5.0.4"
psycopg2 = "^2.9.6"
logging = "^0.4.9.6"
psycopg2-binary = "^2.9.6"
validators = "^0.20.0"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
